ref(crons): Remove frontend references to timeline feature flag
