+++ static/app/views/monitors/overview.tsx
@@ -30,7 +30,6 @@ import usePageFilters from 'sentry/utils/usePageFilters';
 import useRouter from 'sentry/utils/useRouter';
 
 import CronsFeedbackButton from './components/cronsFeedbackButton';
-import {OverviewTable} from './components/overviewTable';
 import {OverviewTimeline} from './components/overviewTimeline';
 import {Monitor} from './types';
 import {makeMonitorListQueryKey} from './utils';
@@ -80,10 +79,6 @@ export default function Monitors() {
     });
   };
 
-  const monitorsTimelineView = organization.features.includes(
-    'crons-timeline-listing-page'
-  );
-
   return (
     <SentryDocumentTitle title={`Crons — ${organization.slug}`}>
       <Layout.Page>
@@ -126,11 +121,7 @@ export default function Monitors() {
               <LoadingIndicator />
             ) : monitorList?.length ? (
               <Fragment>
-                {monitorsTimelineView ? (
-                  <OverviewTimeline monitorList={monitorList} />
-                ) : (
-                  <OverviewTable monitorList={monitorList} />
-                )}
+                <OverviewTimeline monitorList={monitorList} />
                 {monitorListPageLinks && <Pagination pageLinks={monitorListPageLinks} />}
               </Fragment>
             ) : (
